@@ -1,3 +1,3 @@
 module github.com/cornelk/orderedmap
 
-go 1.17
+go 1.24
